update product data structure --progress

+++ src/models/product-types/product-types.service.ts
@@ -3,34 +3,21 @@ import {
   CreateProductTypeDto,
   UpdateProductTypeDto,
 } from '@/models/product-types/dto';
+import { RequestUser } from '@/interfaces';
 import { InjectModel } from '@nestjs/mongoose';
 import { ProductType } from '@/models/product-types/schemas/product-type.schema';
-import mongoose, { Model } from 'mongoose';
+import { Model } from 'mongoose';
 import { ResponseProductTypeDto } from '@/models/product-types/dto/response-product-type.dto';
-import { ProductCollection } from '@/models/product-collections/schemas/product-collection.schema';
-import { SYSTEM } from '@/common/constants';
-import { ResponseBranchDto } from '@/models/branches/dto/response-branch.dto';
+import { ProductsService } from '@/models/products/products.service';
 
 @Injectable()
 export class ProductTypesService {
   constructor(
-    @InjectModel(ProductType.name) private productTypeModel: Model<ProductType>,
-    @InjectModel(ProductCollection.name)
-    private productCollectionModel: Model<ProductCollection>,
+    @InjectModel(ProductType.name)
+    private productTypeModel: Model<ProductType>,
+    private productService: ProductsService,
   ) {}
-
   async create(createProductTypeDto: CreateProductTypeDto, createdBy: string) {
-    const isValidId = mongoose.Types.ObjectId.isValid(
-      createProductTypeDto.parentCollection,
-    );
-    if (!isValidId) throw new Error('Invalid Collection ID');
-
-    const collection = await this.productCollectionModel.findById(
-      createProductTypeDto.parentCollection,
-    );
-
-    if (!collection) throw new Error('Invalid Collection ID');
-
     const newProductType = new this.productTypeModel({
       ...createProductTypeDto,
       createdBy,
@@ -41,12 +28,19 @@ export class ProductTypesService {
     return newProductType.save();
   }
 
-  async findAll() {
-    const productTypes = await this.productTypeModel
-      .find()
-      .populate('parentCollection')
+  async findByName(name: string) {
+    const productType = await this.productTypeModel
+      .findOne({ name })
       .lean()
       .exec();
+
+    if (!productType) throw new Error('Product Type not found');
+
+    return new ResponseProductTypeDto(productType);
+  }
+
+  async findAll() {
+    const productTypes = await this.productTypeModel.find().lean().exec();
 
     if (!productTypes) throw new Error('Product Types not found');
 
@@ -56,11 +50,7 @@ export class ProductTypesService {
   }
 
   async findOne(id: string) {
-    const productType = await this.productTypeModel
-      .findById(id)
-      .populate('parentCollection')
-      .lean()
-      .exec();
+    const productType = await this.productTypeModel.findById(id).lean().exec();
 
     if (!productType) throw new Error('Product Type not found');
 
@@ -72,24 +62,12 @@ export class ProductTypesService {
     updateProductTypeDto: UpdateProductTypeDto,
     updatedBy: string,
   ) {
-    if (updateProductTypeDto.parentCollection) {
-      const isValidId = mongoose.Types.ObjectId.isValid(
-        updateProductTypeDto?.parentCollection,
-      );
-      if (!isValidId) throw new Error('Invalid Collection ID');
-    }
-
     const updatedProductType = await this.productTypeModel
       .findByIdAndUpdate(
         id,
-        {
-          ...updateProductTypeDto,
-          updatedBy,
-          updatedAt: new Date(),
-        },
+        { ...updateProductTypeDto, updatedBy, updatedAt: new Date() },
         { new: true },
       )
-      .populate('parentCollection')
       .lean()
       .exec();
 
@@ -99,6 +77,13 @@ export class ProductTypesService {
   }
 
   async remove(id: string) {
+    const productCount = await this.productService.getProductCountByType(id);
+
+    if (productCount > 0)
+      throw new Error(
+        `Cannot delete product type. ${productCount} product${productCount > 1 ? 's' : ''} assigned to it.`,
+      );
+
     const deletedProductType = await this.productTypeModel
       .findByIdAndDelete(id)
       .lean()
@@ -107,31 +92,5 @@ export class ProductTypesService {
     if (!deletedProductType) throw new Error('Product Type not found');
 
     return new ResponseProductTypeDto(deletedProductType);
-  }
-
-  async getProductTypeCountForCollection(collectionId: string) {
-    return await this.productTypeModel
-      .find({ collection: collectionId })
-      .countDocuments()
-      .exec();
-  }
-
-  async updateProductCount(productTypeId: string, action: 'inc' | 'dec') {
-    const updatedProductType = await this.productTypeModel
-      .findByIdAndUpdate(
-        productTypeId,
-        {
-          updatedAt: new Date(),
-          updatedBy: SYSTEM,
-          $inc: { productCount: action === 'inc' ? 1 : -1 },
-        },
-        { new: true },
-      )
-      .lean()
-      .exec();
-
-    if (!updatedProductType) throw new Error('Product type not found');
-
-    return new ResponseProductTypeDto(updatedProductType);
   }
 }
